sua lai 1 chut

+++ lab2/Bai_tap_3/Sophuc.h
@@ -3,8 +3,8 @@
 
 class Sophuc{
 private:
-    int iThuc;
-    int iAo;
+    double iThuc;
+    double iAo;
 public:
     void Nhap();
     void Xuat();
